--- a/packages/runtime/src/reflection-create.ts
+++ b/packages/runtime/src/reflection-create.ts
@@ -8,6 +8,16 @@ import {MESSAGE_TYPE} from './message-type-contract';
  * information.
  */
 export function reflectionCreate<T extends object>(type: IMessageType<T>): T {
+    /**
+     * This ternary can be removed in the next major version.
+     * The `Object.create()` code path utilizes a new `messagePrototype`
+     * property on the `IMessageType` which has this same `MESSAGE_TYPE`
+     * non-enumerable property on it. Doing it this way means that we only
+     * pay the cost of `Object.defineProperty()` once per `IMessageType`
+     * class of once per "instance". The falsy code path is only provided
+     * for backwards compatibility in cases where the runtime library is
+     * updated without also updating the generated code.
+     */
     const msg: UnknownMessage = type.messagePrototype
         ? Object.create(type.messagePrototype)
         : Object.defineProperty({}, MESSAGE_TYPE, {value: type});