@jcready 2023-12-01
I would greatly appreciate a review of this PR. It's a fairly small change with a big speed improvement. Thank you for your time! 🙏