@timostamm 2023-12-05
Released in [v2.9.2](https://github.com/timostamm/protobuf-ts/releases/tag/v2.9.2).