@Berd74 2024-03-06
Hi,
Is there a way to generate the code in the old way? I mean to generate the objects (create/ function) using:
Object.defineProperty({}, MESSAGE_TYPE, {value: type});
?

Our code base relies on plain objects. Due to the changes, the outcoming objects from the generated code are no longer plain.

We would love to still keep up to date with the updates but right now we are forced to use the old version of the library (2.9.1).