@timostamm 2024-03-07
@jcready, we're using the prototype to avoid Object.defineProperty's perf penalty, while still keeping the property non-enumerable - correct?

@Berd74, I assume the object isn't considered to be plain because of its prototype? You could write a function that strips a message (and all messages it contains in properties) of its prototype. Can you give this a try?