@Berd74 2024-03-08
> @Berd74, I assume the object isn't considered to be plain because of its prototype? You could write a function that strips a message (and all messages it contains in properties) of its prototype. Can you give this a try?

This is our solution right now. After receiving objects from protobuf-ts (in our project: after using services like fetch requests, packets from webSocket, or packets from WebRTC data channels) we use a function that deeply checks if the object prototype has symbol `MESSAGE_TYPE`. If the object meets the condition we convert it back to the objects from version v2.9.1.

before transform:
<img width="681" alt="image" src="https://github.com/timostamm/protobuf-ts/assets/19754650/10bf9145-abcb-4e42-96e9-21492030d9d5">
after transform:
<img width="730" alt="image" src="https://github.com/timostamm/protobuf-ts/assets/19754650/88f47f74-08d4-4cd1-a224-67a9895a01e0">

The solution is not perfect but it works. We hope in the future we will be able to get rid of this transformation.

@jcready I looked at the issue mentioned in [#618](https://github.com/timostamm/protobuf-ts/pull/618) and I can confirm that we are having the same problem. We also use mobx.