@timostamm 2024-03-13
>  I don't believe the issue is related to having this other non-enumerable property on the prototype, just that things like constructor or toString() don't appear as "own properties" of the first prototype in the object's prototype chain or don't match Object.

You're right. Thanks for digging up the details. 

I'm afraid that anything but a plain object (only own properties, no cyclic references, only JSON serializable types) will cause an issue in one framework or the other. At the same time, Protobuf messages ideally have the schema information attached to the object, so all necessary information for serialization and reflection is in one place. It can be a bit frustrating that it‘s extremely rare to see a framework provide hooks so that behavior can be adjusted to the domain models.

https://github.com/timostamm/protobuf-ts/pull/618 isn't going to solve this issue completely, but it does look like an improvement.