@timostamm 2024-03-13
https://github.com/timostamm/protobuf-ts/pull/618 was released in [v2.9.4](https://github.com/timostamm/protobuf-ts/releases/tag/v2.9.4). @Berd74, this should work for you as well.